% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions.R
\name{initial_poll_cleaning}
\alias{initial_poll_cleaning}
\title{Initial Poll Cleaning}
\usage{
initial_poll_cleaning(data)
}
\arguments{
\item{data}{the .csv data}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Takes the 538 polling csv and maks it useable for this repo
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
