% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions.R
\name{probability_dataset}
\alias{probability_dataset}
\title{DataSet for probability_winning_plot function}
\usage{
probability_dataset(data_set)
}
\arguments{
\item{data_set}{The data the user chooses}
}
\value{
polling data
}
\description{
Allows the user to choose between the actual and fake datasets
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
