% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_functions.R
\name{probability_winning_plot}
\alias{probability_winning_plot}
\title{Probability of Winning}
\usage{
probability_winning_plot(candidate)
}
\arguments{
\item{candidate}{Enter the either "Perdue" or "Ossoff"}
}
\value{
Returns a plot
}
\description{
Plots the probabilities for Georgia Senate Runoff candidates David Perdue and Jon Ossoff using bootstrapping methods
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggthemes]{theme_fivethirtyeight}}
}
